Members have valid phone numbers
Caltaiko needs to keep phone numbers in their directory of members to
be used when reaching out to past members. The phonelib gem provides
phone number validation and phone number formatting. Phone numbers can
be added from the new member page and viewed on the database page.

The phone attribute is validated when trying create a member, but could
probably benefit from some Postgres level validations to make sure that
numbers that hit the db are sanitized (i.e. when creating db:seed).

https://github.com/calraijintaiko/caltaiko/issues/41